erdrs: some tests changed to examples
