+++ sdk/erdrs/examples/account.rs
@@ -1,0 +1,17 @@
+use mx_sdk_erdrs::{
+    blockchain::rpc::{ElrondProxy, DEVNET_GATEWAY},
+    data::address::Address,
+};
+
+#[tokio::main]
+async fn main() {
+    let addr = Address::from_bech32_string(
+        "erd1qqqqqqqqqqqqqpgqfzydqmdw7m2vazsp6u5p95yxz76t2p9rd8ss0zp9ts",
+    )
+    .unwrap();
+
+    let blockchain = ElrondProxy::new(DEVNET_GATEWAY.to_string());
+    let account = blockchain.get_account(&addr).await.unwrap();
+
+    println!("account: {:#?}", account);
+}
