+++ sdk/erdrs/examples/tx_default_args.rs
@@ -1,0 +1,21 @@
+use mx_sdk_erdrs::{
+    blockchain::rpc::{ElrondProxy, DEVNET_GATEWAY},
+    data::address::Address,
+};
+
+#[tokio::main]
+async fn main() {
+    let blockchain = ElrondProxy::new(DEVNET_GATEWAY.to_string());
+    let network_config = blockchain.get_network_config().await.unwrap();
+    let addr = Address::from_bech32_string(
+        "erd1qqqqqqqqqqqqqpgqfzydqmdw7m2vazsp6u5p95yxz76t2p9rd8ss0zp9ts",
+    )
+    .unwrap();
+
+    let arg = blockchain
+        .get_default_transaction_arguments(&addr, &network_config)
+        .await
+        .unwrap();
+
+    println!("default tx arg: {:#?}", arg);
+}
