+++ sdk/erdrs/src/blockchain/mod.rs
@@ -1,4 +1,1 @@
 pub mod rpc;
-
-#[cfg(test)]
-mod tests;
