+++ sdk/erdrs/src/interactors/mod.rs
@@ -1,4 +1,1 @@
-#[cfg(test)]
-mod wallet_test;
-
 pub mod wallet;
